worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "trio>=0.30.0",
]
